requirements.txt
Django>=5.1,<5.2
python-dotenv
gunicorn
psycopg2-binary
dj-database-url
requests
beautifulsoup4
whitenoise
